delete base components replace cell with enhanced input component
