--- src/components/base/index.js
@@ -1,3 +1,0 @@
-export { default as InputButton } from './InputButton';
-export { default as InputField } from './InputField';
-export { default as Select } from './Select';
